Cleaned up the DeepZipper source file.
- Minor method refactorings on DeepZipper.
- Took out PathCreator and BasicMergeStrategy into separate files.
