+++ src/test/scala/com/codecommit/antixml/DeepZipperSpecs.scala
@@ -3,19 +3,18 @@ package com.codecommit.antixml
 import org.specs2.mutable._
 import org.specs2.ScalaCheck
 import org.scalacheck._
-import org.specs2.matcher.DataTables
 import XML._
 import com.codecommit.antixml.DeepZipper._
 import scala.io.Source
 
-class DeepZipperSpecs extends SpecificationWithJUnit with ScalaCheck with DataTables with XMLGenerators {
+class DeepZipperSpecs extends SpecificationWithJUnit with ScalaCheck  with XMLGenerators {
 
   implicit val params = set(maxSize -> 10, minTestsOk -> 20)
   val bookstore = resource("bookstore.xml")
   val onlyBell = Group(<bookstore><book><title>For Whom the Bell Tolls</title><author>Hemmingway</author></book></bookstore>.convert)
   val onlyPS = Group(<bookstore><book><title>Programming Scala</title><author>Dean Wampler</author><author>Alex Payne</author></book></bookstore>.convert)
   
-  // This is by no means exhaustive, just showing off what's possible.
+  // TODO This is by no means exhaustive, just showing off what's possible.
 
   "Deep Zipper selection" should {
     // I'm lazy so the bookstore is hardcoded here.
@@ -490,169 +489,6 @@ class DeepZipperSpecs extends SpecificationWithJUnit with ScalaCheck with DataTa
       val ns = fromString("<parent>Some text<sub1><target>sub1</target></sub1><target>top<sub1><target>top1</target><target>top2</target></sub1><target>top3-outer</target></target><phoney><target>phoney</target></phoney>More text<target>outside</target></parent>")
       val result = fromString("<parent><target>top<sub1><target>top1</target><target>top2</target></sub1><target>top3-outer</target></target><target>outside</target><target>sub1</target><target>top3-outer</target><target>phoney</target><target>top1</target><target>top2</target></parent>")
       ns ~ "target" mustEqual result.children
-    }
-  }
-
-  "A basic merge startegy" should {
-    def maxTime(s: Seq[(Node, Int)]) = s.maxBy(_._2)
-
-    "fail on empty lists" in {
-      val n = Text("foo")
-      BasicNodeMergeStrategy(n, List[(Node, Int)]()) must throwA[Exception]
-    }
-
-    implicit def timeNodes = Arbitrary { for { n <- nodeGenerator(1); t <- Gen.posNum[Int] } yield (n, t) }
-
-    "select the newest node when the original is not an elem" in check { alts: List[(Node, Int)] =>
-      !alts.isEmpty ==> {
-        (BasicNodeMergeStrategy(_: Node, alts) mustEqual maxTime(alts)).forall {
-          Seq(ProcInstr("foo", "bar"), Text("baz"), CDATA("qux"), EntityRef("somethin"))
-        }
-      }
-    }
-
-    "select the newest node when no elems are present in the alternatives" in check { alts: List[(Node, Int)] =>
-      val noElems = alts.filterNot(_._1.isInstanceOf[Elem])
-      !noElems.isEmpty ==> {
-        BasicNodeMergeStrategy(elem("fo0"), noElems) mustEqual noElems.maxBy(_._2)
-      }
-    }
-
-    val orig = elem("orig", Text("foo"))
-    val alt1 = (orig.copy(name = "alt1"), 2)
-    val alt2 = (orig.copy(name = "alt2"), 5)
-    val alt3 = (orig.copy(name = "orig", children = Group(Text("bar"), Text("baz"))), 4)
-    val alt4 = (orig.copy(name = "alt4", children = Group()), 3)
-    val text = (Text("baz"), 3)
-
-    "merge elems by taking the most recent children and most recent non child modification" in {
-      val alts1 = Seq(alt1, alt2, text, alt3, alt4)
-      BasicNodeMergeStrategy(orig, alts1) mustEqual (orig.copy(name = alt2._1.name, children = alt3._1.children), 5)
-    }
-
-    "keep the children intact if they were not modified" in {
-      val alts2 = Seq(alt1, alt2, text)
-      BasicNodeMergeStrategy(orig, alts2) mustEqual (orig.copy(name = alt2._1.name), 5)
-    }
-
-    "keep non children fields intact if they were not modified" in {
-      val alts3 = Seq(text, alt3, alt4)
-      BasicNodeMergeStrategy(orig, alts3) mustEqual (orig.copy(children = alt3._1.children), 4)
-    }
-  }
-
-  "Path values" should {
-    import PathCreator._
-
-    val s = *
-
-    "ignore empty groups" in {
-      val empty = Group()
-
-      fromNodes(s)(empty) mustEqual Nil
-      all(s)(empty) mustEqual Nil
-      directChildren(s)(empty) mustEqual Nil
-      allChildren(s)(empty) mustEqual Nil
-    }
-
-    val x0 = fromString("<root0><a0>foo</a0><b0>baz</b0><c0/></root0>")
-    val a0 = fromString("<a0>foo</a0>"); val b0 = fromString("<b0>baz</b0>"); val c0 = fromString("<c0/>")
-    val foo0 = Text("foo"); val baz0 = Text("baz")
-
-    val x1 = fromString("<root1><a1>foo</a1><b1>baz</b1><c1/></root1>")
-    val a1 = fromString("<a1>foo</a1>"); val b1 = fromString("<b1>baz</b1>"); val c1 = fromString("<c1/>")
-    val foo1 = Text("foo"); val baz1 = Text("baz")
-
-    val x2 = fromString("<root2><a2>foo</a2><b2>baz</b2><c2/></root2>")
-    val a2 = fromString("<a2>foo</a2>"); val b2 = fromString("<b2>baz</b2>"); val c2 = fromString("<c2/>")
-    val foo2 = Text("foo"); val baz2 = Text("baz")
-
-    val group = Group(x0, x1, x2)
-
-    def ps(pars: (Elem, Int)*) = List(pars.map(ParentLoc.tupled): _*)
-    def nl(n: Node, l: Int) = WithLoc(n, l)
-
-    val root = List((nl(x0, 0), ps()), (nl(x1, 1), ps()), (nl(x2, 2), ps()))
-    val directChild = List(
-      (nl(a0, 0), ps((x0, 0))), (nl(b0, 1), ps((x0, 0))), (nl(c0, 2), ps((x0, 0))),
-      (nl(a1, 0), ps((x1, 1))), (nl(b1, 1), ps((x1, 1))), (nl(c1, 2), ps((x1, 1))),
-      (nl(a2, 0), ps((x2, 2))), (nl(b2, 1), ps((x2, 2))), (nl(c2, 2), ps((x2, 2)))
-    )
-    val rest = List(
-      (nl(foo0, 0), ps((a0, 0), (x0, 0))), (nl(baz0, 0), ps((b0, 1), (x0, 0))),
-      (nl(foo1, 0), ps((a1, 0), (x1, 1))), (nl(baz1, 0), ps((b1, 1), (x1, 1))),
-      (nl(foo2, 0), ps((a2, 0), (x2, 2))), (nl(baz2, 0), ps((b2, 1), (x2, 2)))
-    )
-
-    "take from the root of the nodes" in {
-      fromNodes(s)(group) mustEqual root
-    }
-
-    "take the children of the root nodes" in {
-      directChildren(s)(group) mustEqual directChild
-    }
-
-    "take all the nodes recursively, breadth first" in {
-      all(s)(group) mustEqual root ::: directChild ::: rest
-    }
-
-    "take all children nodes recursively, breadth first" in {
-      allChildren(s)(group) mustEqual directChild ::: rest
-    }
-
-    "apply selectors at the root level" in {
-      val sel = Selector({ case Elem(_, "root1", _, _, _) => elem("selected") })
-      fromNodes(sel)(group) mustEqual List((nl(elem("selected"), 1), ps()))
-    }
-
-    "apply selectors to the children of the root" in {
-      val sel = Selector({ case Elem(_, "b2", _, _, _) => elem("selected") })
-      directChildren(sel)(group) mustEqual List((nl(elem("selected"), 1), ps((x2, 2))))
-    }
-
-    val selDeep = Selector({
-      case Elem(_, "root2", _, _, _) => elem("selected")
-      case Elem(_, "c1", _, _, _) => elem("selected")
-      case Elem(_, "b2", _, _, _) => elem("selected")
-      case Text("baz") => Text("selected")
-    })
-
-    val selResRoot = List(
-      (nl(elem("selected"), 2), ps()))
-
-    val selResNoRoot = List(
-      (nl(elem("selected"), 2), ps((x1, 1))),
-      (nl(elem("selected"), 1), ps((x2, 2))),
-      (nl(Text("selected"), 0), ps((b0, 1), (x0, 0))),
-      (nl(Text("selected"), 0), ps((b1, 1), (x1, 1))),
-      (nl(Text("selected"), 0), ps((b2, 1), (x2, 2)))
-    )
-
-    "apply selectors recursively" in {
-      all(selDeep)(group) mustEqual selResRoot ::: selResNoRoot
-    }
-
-    "apply selectors recursively on the children" in {
-      allChildren(selDeep)(group) mustEqual selResNoRoot
-    }
-  }
-  
-  "Paths" should {
-    "not contain duplicate locations" in {
-      new Path(Seq((WithLoc("foo", 1), Nil), (WithLoc("bar", 1), Nil))) must throwA[IllegalArgumentException]
-    }
-
-    "properly split locations and contents" in {
-      val p1 = ParentLoc(elem("a"), 1) :: Nil
-      val loc1 = WithLoc("foo", 2)
-      val p2 = ParentLoc(elem("b"), 1) :: ParentLoc(elem("c"), 2) :: Nil
-      val loc2 = WithLoc("bar", 2)
-      val path = new Path(Seq((loc1, p1), (loc2, p2)))
-
-      path.contents mustEqual Seq("foo", "bar")
-      path.locs mustEqual Seq(
-        LocationContext(loc1.loc, p1, DeepZipper.initTime),
-        LocationContext(loc2.loc, p2, DeepZipper.initTime))
     }
   }
